Lower warning on help.smile.io cert check.
Helpscout uses lets-encrypt certs, and renews them automatically when
they're due to expire in 30 days. Because of this our previous setting
of 60 days was causing false positives. Bumping the warning level down
to 25 days still gives us lots of advance warning of something is wrong
but gives Helpscout time to renew the cert automatically as well.